backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
geoalchemy2
psycopg2-binary
pydantic-settings
redis
